refactor: deduplicate and split models
